+++ sema4ai/vscode-client/src/extension.ts
@@ -156,6 +156,7 @@ import {
     SEMA4AI_AGENT_PACKAGE_PUBLISH_TO_SEMA4_AI_STUDIO_APP,
     SEMA4AI_AGENT_PACKAGE_IMPORT,
     SEMA4AI_UPDATE_AGENT_VERSION,
+    SEMA4AI_REFRESH_AGENT_SPEC,
 } from "./robocorpCommands";
 import { installWorkspaceWatcher } from "./pythonExtIntegration";
 import { refreshCloudTreeView } from "./viewsRobocorp";
@@ -189,6 +190,7 @@ import {
     packAgentPackage,
     importAgentPackage,
     updateAgentVersion,
+    refreshAgentSpec,
 } from "./robo/agentPackage";
 import { getSema4AIStudioURLForAgentZipPath, getSema4AIStudioURLForFolderPath } from "./deepLink";
 import { LocalPackageMetadataInfo } from "./protocols";
@@ -498,6 +500,7 @@ function registerRobocorpCodeCommands(C: CommandRegistry, context: ExtensionCont
     C.register(SEMA4AI_PACK_AGENT_PACKAGE, async (agentPath: string) => selectAndPackAgentPackage(agentPath));
     C.register(SEMA4AI_OPEN_RUNBOOK_TREE_SELECTION, (robot: RobotEntry) => views.openRunbookTreeSelection(robot));
     C.register(SEMA4AI_UPDATE_AGENT_VERSION, async (agentPath: string) => updateAgentVersion(agentPath));
+    C.register(SEMA4AI_REFRESH_AGENT_SPEC, async (agentPath: string) => refreshAgentSpec(agentPath));
 }
 
 async function clearEnvAndRestart() {